scripts/seo_indexing_agent_requirements.txt
google-api-python-client==2.158.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
requests==2.32.3
beautifulsoup4==4.12.3
